requirements.txt
faiss_cpu==1.7.4
fastapi==0.109.2
llama_index==0.10.6
pydantic==2.6.1
uvicorn==0.27.1
